@@ -1,8 +1,9 @@
 {
   "name": "schmucklicloud_auth",
-  "version": "1.4.0",
+  "version": "1.5.0",
   "description": "This is the official SDK for schmuckliCloud Authentication service for JavaScript.",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -23,18 +24,18 @@
   },
   "homepage": "https://schmuckli.cloud",
   "dependencies": {
-    "axios": "^0.19.0"
+    "axios": "^0.19.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/register": "^7.5.5",
+    "@babel/core": "^7.11.6",
+    "@babel/register": "^7.11.5",
     "babel-core": "^6.26.3",
-    "babel-loader": "^8.0.6",
+    "babel-loader": "^8.1.0",
     "babel-polyfill": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
-    "mocha": "^6.2.0",
-    "webpack": "^4.39.2",
-    "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^3.8.0"
+    "mocha": "^6.2.3",
+    "webpack": "^4.44.2",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0"
   }
 }
